--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "pybind11"]
-	path = pybind11
-	url = https://github.com/pybind/pybind11
