+++ src/mobase/wrappers/pyfiletree.cpp
@@ -3,12 +3,7 @@
 #include <tuple>
 #include <variant>
 
-#include <pybind11/functional.h>
-#include <pybind11/operators.h>
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-
-#include "../pybind11_qt/pybind11_qt.h"
+#include "../pybind11_all.h"
 
 #include <ifiletree.h>
 #include <log.h>
