+++ src/proxy/plugin_python_en.ts
@@ -4,58 +4,58 @@
 <context>
     <name>ProxyPython</name>
     <message>
-        <location filename="proxypython.cpp" line="162"/>
+        <location filename="proxypython.cpp" line="166"/>
         <source>Python Proxy</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="172"/>
+        <location filename="proxypython.cpp" line="176"/>
         <source>Proxy Plugin to allow plugins written in python to be loaded</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="244"/>
+        <location filename="proxypython.cpp" line="248"/>
         <source>ModOrganizer path contains a semicolon</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="246"/>
+        <location filename="proxypython.cpp" line="250"/>
         <source>Python DLL not found</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="248"/>
+        <location filename="proxypython.cpp" line="252"/>
         <source>Invalid Python DLL</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="250"/>
+        <location filename="proxypython.cpp" line="254"/>
         <source>Initializing Python failed</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="252"/>
-        <location filename="proxypython.cpp" line="282"/>
+        <location filename="proxypython.cpp" line="256"/>
+        <location filename="proxypython.cpp" line="286"/>
         <source>invalid problem key %1</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="260"/>
+        <location filename="proxypython.cpp" line="264"/>
         <source>The path to Mod Organizer (%1) contains a semicolon. &lt;br&gt;While this is legal on NTFS drives, many softwares do not handle it correctly.&lt;br&gt;Unfortunately MO depends on libraries that seem to fall into that group.&lt;br&gt;As a result the python plugin cannot be loaded, and the only solution we canoffer is to remove the semicolon or move MO to a path without a semicolon.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="271"/>
+        <location filename="proxypython.cpp" line="275"/>
         <source>The Python plugin DLL was not found, maybe your antivirus deleted it. Re-installing MO2 might fix the problem.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="274"/>
+        <location filename="proxypython.cpp" line="278"/>
         <source>The Python plugin DLL is invalid, maybe your antivirus is blocking it. Re-installing MO2 and adding exclusions for it to your AV might fix the problem.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="proxypython.cpp" line="279"/>
+        <location filename="proxypython.cpp" line="283"/>
         <source>The initialization of the Python plugin DLL failed, unfortunately without any details.</source>
         <translation type="unfinished"></translation>
     </message>
@@ -63,12 +63,7 @@
 <context>
     <name>QObject</name>
     <message>
-        <location filename="../runner/error.h" line="68"/>
-        <source>An unexpected C++ exception was thrown in python code.</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../runner/error.h" line="117"/>
+        <location filename="../runner/error.h" line="75"/>
         <source>An unknown exception was thrown in python code.</source>
         <translation type="unfinished"></translation>
     </message>
